test: api 테스트 추가, CRUD 테스트 분리, 패키지 분리
