+++ src/test/java/roomescape/AdminPageTest.java
@@ -1,0 +1,48 @@
+package roomescape;
+
+import io.restassured.RestAssured;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.annotation.DirtiesContext;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
+@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
+public class AdminPageTest {
+
+    @Test
+    @DisplayName("200 상태코드와 관리자 페이지를 반환한다.")
+    void returnAdminPageWithStatus200Test() {
+        RestAssured.given().log().all()
+                .when().get("/admin")
+                .then().log().all()
+                .statusCode(200);
+    }
+
+    @Test
+    @DisplayName("200 상태코드와 예약 페이지를 반환한다")
+    void returnReservationPageWithStatus200Test() {
+        RestAssured.given().log().all()
+                .when().get("/admin/reservation")
+                .then().log().all()
+                .statusCode(200);
+    }
+
+    @Test
+    @DisplayName("200 상태코드와 예약 시간 페이지를 반환한다")
+    void returnReservationTimePageWithStatus200Test() {
+        RestAssured.given().log().all()
+                .when().get("/admin/time")
+                .then().log().all()
+                .statusCode(200);
+    }
+
+    @Test
+    @DisplayName("200 상태코드와 테마 페이지를 반환한다")
+    void returnThemePageWithStatus200Test() {
+        RestAssured.given().log().all()
+                .when().get("/admin/theme")
+                .then().log().all()
+                .statusCode(200);
+    }
+}
